fix: route permision

+++ routes/web.php
@@ -187,13 +187,13 @@ Route::group(['middleware' => ['auth:hr']], function () {
     Route::get('/kurikulum/show/{id}', [KurikulumController::class, 'show'])->name('kurikulum.show')->middleware(['permission:read_kurikulum']);
 
     //kelas
-    Route::get('/kelas', [KelasController::class, 'index'])->name('kelas.index');
-    Route::get('/kelas/create', [KelasController::class, 'create'])->name('kelas.create');
-    Route::post('/kelas', [KelasController::class, 'store'])->name('kelas.store');
-    Route::get('/kelas/{kelas}/edit', [KelasController::class, 'edit'])->name('kelas.edit');
-    Route::put('/kelas/{kelas}', [KelasController::class, 'update'])->name('kelas.update');
-    Route::delete('/kelas/{kelas}', [KelasController::class, 'destroy'])->name('kelas.destroy');
-    Route::get('/kelas/show/{id}', [KelasController::class, 'show'])->name('kelas.show');
+    Route::get('/kelas', [KelasController::class, 'index'])->name('kelas.index')->middleware(['permission:read_kelas']);
+    Route::get('/kelas/create', [KelasController::class, 'create'])->name('kelas.create')->middleware(['permission:create_kelas']);
+    Route::post('/kelas', [KelasController::class, 'store'])->name('kelas.store')->middleware(['permission:create_kelas']);
+    Route::get('/kelas/{kelas}/edit', [KelasController::class, 'edit'])->name('kelas.edit')->middleware(['permission:update_kelas']);
+    Route::put('/kelas/{kelas}', [KelasController::class, 'update'])->name('kelas.update')->middleware(['permission:update_kelas']);
+    Route::delete('/kelas/{kelas}', [KelasController::class, 'destroy'])->name('kelas.destroy')->middleware(['permission:delete_kelas']);
+    Route::get('/kelas/show/{id}', [KelasController::class, 'show'])->name('kelas.show')->middleware(['permission:read_kelas']);
     Route::get('/kelas/details/{kurikulum}', [KelasController::class, 'getKurikulumDetails'])->name('kelas.details');
 
     //periode perkuliahan
@@ -220,7 +220,7 @@ Route::group(['middleware' => ['auth:hr']], function () {
     Route::put('/skala-nilai/{skalaNilai}', [SkalaNilaiController::class, 'update'])->name('skala-nilai.update');
     Route::delete('/skala-nilai/{skalaNilai}', [SkalaNilaiController::class, 'destroy'])->name('skala-nilai.destroy');
     Route::get('/skala-nilai/show/{id}', [SkalaNilaiController::class, 'show'])->name('skala-nilai.show');
-    
+
 });
 
 //login
